requirements.txt
openai>=1.50.0
yt-dlp>=2024.10.0
faster-whisper>=1.0.3
imagehash>=4.3.1
Pillow>=10.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
tqdm>=4.66.0
